requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
pgvector>=0.3.0
openai>=1.35.0
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-multipart>=0.0.9
pypdf>=4.2.0
python-docx>=1.1.2
redis>=5.0.7
pytest>=8.2.2
httpx>=0.27.0
pytest-asyncio>=0.23.7
